refactored_ui_slice
