+++ Clients/src/application/redux/store.ts
@@ -1,7 +1,7 @@
 import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
-import uiSlice from "../../presentation/tools/uiSlice";
+import uiSlice from "./ui/uiSlice";
 import authReducer from "./auth/authSlice";
 import fileReducer from "./file/fileSlice";
