+++ Clients/src/application/repository/annex_struct_iso.repository.ts
@@ -1,6 +1,6 @@
 import { GetRequestParams } from "../../domain/interfaces/iRequestParams";
 import { apiServices } from "../../infrastructure/api/networkServices";
-import { getAuthToken } from "../redux/getAuthToken";
+import { getAuthToken } from "../redux/auth/getAuthToken";
 
 export async function GetAllAnnexes({
   routeUrl,
